vlm_service/requirements.txt
redis
pillow
python-dotenv
transformers>=4.48.0
accelerate>=0.26.0
sentencepiece
einops